NetworkFilterList has been moved to network_filter_list.rs
